Improve tests on pagination
